@@ -21,7 +21,7 @@
   "classmap-authoritative": true,
   "prefer-stable": true,
   "require": {
-    "php": ">=7.3",
+    "php": ">=7.4",
     "laminas/laminas-session": "2.10.*",
     "laminas/laminas-cache": "2.11.*",
     "laminas/laminas-serializer": "2.10.*",
